style(grid): ui improvements
Created dynamic grid header item / cell component.
Created custom checkbox style.
