+++ projects/data-grid/src/lib/components/column-selector/column-selector.component.html
@@ -2,10 +2,15 @@
   {{localeTexts.title | ngrxTranslate}} <i class="arrow down"></i>
 </button>
 
-<div id="checkboxes" *ngIf="expanded">
-  <label *ngFor="let column of columns; trackBy: trackByIndex; let i = index" [for]="getHeaderName(column)">
-    <input type="checkbox" [id]="getColumnId(column)" [checked]="getColumnVisible(column)"
-           (change)="onToggleColumn(column)"/>
-    {{getHeaderName(column)}}
-  </label>
+<div *ngIf="expanded" id="checkboxes">
+  <div  class="data-grid-checkbox-wrapper data-grid-checkbox" *ngFor="let column of columns; trackBy: trackByIndex; let i = index">
+      <input type="checkbox" [id]="getColumnId(column)"
+             [checked]="getColumnVisible(column)"
+             (change)="onToggleColumn(column)"
+             class="data-grid-input"/>
+      <label  [for]="getColumnId(column)"
+              class="data-grid-label">
+        {{getHeaderName(column)}}
+      </label>
+  </div>
 </div>
