+++ projects/data-grid/src/lib/components/column-selector/column-selector.component.scss
@@ -1,26 +1,23 @@
 @import "../../../styles/variables";
 
-.selectBox {
-  position: relative;
-}
+:host {
+  margin: $size-s 0 $size-s auto;
 
-#checkboxes {
-  position: absolute;
-  z-index: 3;
-  background: #fff;
-  border: 1px #dadada solid;
-}
-
-#checkboxes label {
-  padding: 10px 25px;
-  display: block;
-  color: $text-color;
-
-  input {
-    margin-right: $size-xs;
+  .selectBox {
+    position: relative;
   }
+
+  #checkboxes {
+    position: absolute;
+    z-index: 3;
+    background: #fff;
+    border: 1px #dadada solid;
+    right: 0;
+    left: auto;
+    min-width: 142px;
+    padding: 0 $size-s;
+  }
+
 }
 
-#checkboxes label:hover {
-  background-color: $border-color;
-}
+
