+++ projects/data-grid/src/lib/components/filter/filter.component.scss
@@ -8,6 +8,7 @@
   right: -20px;
   width: 300px;
   border: 1px solid $border-color;
+  z-index: 2;
 }
 
 .button-calendar {
