+++ projects/data-grid/src/lib/components/grid-display.component.html
@@ -7,6 +7,7 @@
       [isSelected]="checkSelected(i)"
       [columns]="columns"
       [data]="rowData"
+      [rowIndex]="i"
       (toggleRow)="onToggleRow(i)">
     </ngrx-grid-row>
   </div>
