+++ projects/data-grid/src/lib/components/grid-header/dynamic-grid-header-item.component.html
@@ -1,4 +1,4 @@
-<h5 [title]="headerName | uppercase">{{headerName | uppercase}} </h5>
+<h5 [title]="headerName">{{headerName | uppercase}} </h5>
 
 <button *ngIf="sortAvailable" class="sort-btn" (click)="onApplySort()">
   <i class="caret caret-up" [ngStyle]="upCaretStyle"></i>
