+++ projects/data-grid/src/lib/components/grid-header/grid-header.component.html
@@ -1,16 +1,22 @@
-<div class="grid-checkbox-header">
-  <input
-    *ngIf="checkboxSelection"
-    type="checkbox"
-    [checked]="allRowsSelected"
-    (change)="onToggleSelectAllRows()"/>
-</div>
+<ngrx-grid-header-item>
+  <div class="data-grid-checkbox-wrapper data-grid-checkbox">
+    <input
+      *ngIf="checkboxSelection"
+      type="checkbox"
+      [checked]="allRowsSelected"
+      (change)="onToggleSelectAllRows()"
+      id="select-all"
+      class="data-grid-input" />
+    <label class="data-grid-label" for="select-all"></label>
+  </div>
+</ngrx-grid-header-item>
 
 <ng-container *ngFor="let column of columns; trackBy: trackByIndex; let i = index">
-  <ngrx-grid-header-item
-    *ngIf="column.visible"
-    [column]="column"
-    (filterGrid)="filterGrid.emit($event)"
-    (sortGrid)="sortGrid.emit($event)">
+  <ngrx-grid-header-item *ngIf="column.visible">
+    <ngrx-dynamic-grid-header-item
+      [column]="column"
+      (filterGrid)="filterGrid.emit($event)"
+      (sortGrid)="sortGrid.emit($event)">
+    </ngrx-dynamic-grid-header-item>
   </ngrx-grid-header-item>
 </ng-container>
