+++ projects/data-grid/src/lib/components/grid-header/grid-header.component.scss
@@ -5,7 +5,6 @@
 }
 
 .grid-checkbox-header {
-  background-color: $header-color;
   height: $header-height;
   display: flex;
   justify-content: center;
