+++ projects/data-grid/src/lib/components/grid-row/grid-cell/grid-cell.component.ts
@@ -1,37 +1,11 @@
-import {
-  Component,
-  ComponentFactoryResolver,
-  Input,
-  OnInit,
-  ViewChild,
-  ViewContainerRef
-} from '@angular/core';
-import { GridCellDirective } from '../../../directives/grid-cell.directive';
-import { DataGridColumnWithId, GridCell } from '../../../models';
-import { DefaultGridCellComponent } from './default-grid-cell.component';
+import { Component } from '@angular/core';
 
 @Component({
   selector: 'ngrx-grid-cell',
-  templateUrl: 'grid-cell.component.html',
-  styleUrls: ['grid-cell.component.scss']
+  template: '<ng-content></ng-content>',
+  styleUrls: ['./grid-cell.component.scss']
 })
-export class GridCellComponent implements OnInit {
-  @Input() data: any;
-  @Input() column: DataGridColumnWithId;
+export class GridCellComponent {
 
-  @ViewChild(GridCellDirective, {read: ViewContainerRef, static: true}) cellHost: ViewContainerRef;
 
-  constructor(private componentFactoryResolver: ComponentFactoryResolver) {}
-
-  ngOnInit(): void {
-    this.cellHost.clear();
-    this.loadComponent();
-  }
-
-  private loadComponent(): void {
-    const cellComponent = this.column.component || DefaultGridCellComponent;
-    const componentFactory = this.componentFactoryResolver.resolveComponentFactory<GridCell>(cellComponent);
-    const componentRef = this.cellHost.createComponent(componentFactory);
-    componentRef.instance.data = this.data;
-  }
 }
