+++ projects/data-grid/src/lib/data-grid.module.ts
@@ -4,7 +4,7 @@ import { ReactiveFormsModule } from '@angular/forms';
 import { NgbModule } from '@ng-bootstrap/ng-bootstrap';
 import { GridDisplayComponent } from './components/grid-display.component';
 import { GridRowComponent } from './components/grid-row/grid-row.component';
-import { GridCellComponent } from './components/grid-row/grid-cell/grid-cell.component';
+import { DynamicGridCellComponent } from './components/grid-row/grid-cell/dynamic-grid-cell.component';
 import { ColumnSelectorComponent } from './components/column-selector/column-selector.component';
 import { PaginationComponent } from './components/pagination/pagination.component';
 import { FilterComponent } from './components/filter/filter.component';
@@ -17,6 +17,8 @@ import { GridFooterComponent } from './components/grid-footer/grid-footer.compon
 import { GridStoreConfig, InternalGridStoreConfig, NgrxGridConfig } from './config';
 import { DefaultGridCellComponent } from './components/grid-row/grid-cell/default-grid-cell.component';
 import { TranslatePipe } from './pipes/translate.pipe';
+import { GridCellComponent } from './components/grid-row/grid-cell/grid-cell.component';
+import { DynamicGridHeaderItemComponent } from './components/grid-header/dynamic-grid-header-item.component';
 
 export const DEFAULT_GRID_FEATURE_NAME = 'grid';
 
@@ -33,12 +35,14 @@ export function createDefaultGridStoreConfig(config: NgrxGridConfig): NgrxGridCo
     DataGridComponent,
     GridCellDirective,
     GridRowComponent,
+    DynamicGridCellComponent,
     GridCellComponent,
     ColumnSelectorComponent,
     PaginationComponent,
     FilterComponent,
     GridHeaderComponent,
     GridHeaderItemComponent,
+    DynamicGridHeaderItemComponent,
     GridFooterComponent,
     DefaultGridCellComponent,
     TranslatePipe
