+++ projects/data-grid/src/styles.scss
@@ -1,5 +1,6 @@
 @import "./styles/variables";
 @import "./styles/buttons";
+@import "./styles/checkbox";
 @import "./styles/variables";
 @import url('https://fonts.googleapis.com/css?family=Open+Sans:400,600&display=swap');
 
@@ -19,4 +20,5 @@ body {
   margin: 0;
   padding: 0;
   font-size: 1rem;
+  color: $text-color;
 }
