+++ projects/data-grid/src/styles/_buttons.scss
@@ -66,6 +66,7 @@ i.arrow {
   justify-content: center;
   width: 20px;
   height: 30px;
+  margin-right: 10px;
   border: 0;
   padding: 0;
 
@@ -99,7 +100,7 @@ i.arrow {
   .bar {
     width: 15px;
     height: 2px;
-    background-color: $text-color;
+    background-color: #a0abbb;
     display: block;
     margin: 2px 0;
