+++ projects/data-grid/src/styles/_variables.scss
@@ -1,9 +1,9 @@
 /* COLORS */
 $c_primary: #2659C5;
 $c_primary_faded: rgba(38, 89, 197, 0.1);
-$border-color: #E6ECF8;
+$border-color: #ced4da;
 $border-color-hover: #102E54;
-$text-color: #102E54;
+$text-color: #4b5154;
 $header-color: #F7F8F9;
 $text-color-light: #9FB0C5;
 
@@ -22,6 +22,28 @@ $size-xl: 25px;
 $size-xxl: 30px;
 
 /* Grid Variables */
-$header-height: 50px;
-$row-height: 60px;
+$header-height: 40px;
+$row-height: 40px;
 $cell-padding: 0 $size-l;
+
+
+
+
+/* Checbox */
+
+// Replace `$search` with `$replace` in `$string`
+// Used on our SVG icon backgrounds for custom forms.
+// @author Hugo Giraudel
+@function str-replace($string, $search, $replace: "") {
+  $index: str-index($string, $search);
+
+  @if $index {
+    @return str-slice($string, 1, $index - 1) + $replace + str-replace(str-slice($string, $index + str-length($search)), $search, $replace);
+  }
+
+  @return $string;
+}
+
+$custom-control-indicator-color: #ffffff;
+$data-grid-checkbox-indicator-icon-checked: str-replace(url("data:image/svg+xml,%3csvg xmlns='http://www.w3.org/2000/svg' viewBox='0 0 8 8'%3e%3cpath fill='#{$custom-control-indicator-color}' d='M6.564.75l-3.59 3.612-1.538-1.55L0 4.26 2.974 7.25 8 2.193z'/%3e%3c/svg%3e"), "#", "%23") !default;
+$data-grid-checkbox-indicator-icon-indeterminate: str-replace(url("data:image/svg+xml,%3csvg xmlns='http://www.w3.org/2000/svg' viewBox='0 0 4 4'%3e%3cpath stroke='#{$custom-control-indicator-color}' d='M0 2h4'/%3e%3c/svg%3e"), "#", "%23") !default;
